// src/rdb/models/SiteAllPackagasetsModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ALTRepo API
 * altrepo API v1
 *
 * The version of the OpenAPI document: 1.19.15
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SiteAllPackagasetsElementModel } from './SiteAllPackagasetsElementModel';
import {
    SiteAllPackagasetsElementModelFromJSON,
    SiteAllPackagasetsElementModelFromJSONTyped,
    SiteAllPackagasetsElementModelToJSON,
} from './SiteAllPackagasetsElementModel';

/**
 * 
 * @export
 * @interface SiteAllPackagasetsModel
 */
export interface SiteAllPackagasetsModel {
    /**
     * number of package sets found
     * @type {number}
     * @memberof SiteAllPackagasetsModel
     */
    length?: number;
    /**
     * list of package sets with source package count
     * @type {Array<SiteAllPackagasetsElementModel>}
     * @memberof SiteAllPackagasetsModel
     */
    branches?: Array<SiteAllPackagasetsElementModel>;
}

/**
 * Check if a given object implements the SiteAllPackagasetsModel interface.
 */
export function instanceOfSiteAllPackagasetsModel(value: object): value is SiteAllPackagasetsModel {
    return true;
}

export function SiteAllPackagasetsModelFromJSON(json: any): SiteAllPackagasetsModel {
    return SiteAllPackagasetsModelFromJSONTyped(json, false);
}

export function SiteAllPackagasetsModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): SiteAllPackagasetsModel {
    if (json == null) {
        return json;
    }
    return {
        
        'length': json['length'] == null ? undefined : json['length'],
        'branches': json['branches'] == null ? undefined : ((json['branches'] as Array<any>).map(SiteAllPackagasetsElementModelFromJSON)),
    };
}

export function SiteAllPackagasetsModelToJSON(value?: SiteAllPackagasetsModel | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'length': value['length'],
        'branches': value['branches'] == null ? undefined : ((value['branches'] as Array<any>).map(SiteAllPackagasetsElementModelToJSON)),
    };
}