// src/rdb/models/SiteSourcePackagesVersionsModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ALTRepo API
 * altrepo API v1
 *
 * The version of the OpenAPI document: 1.19.15
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SitePackageVersionsElementModel } from './SitePackageVersionsElementModel';
import {
    SitePackageVersionsElementModelFromJSON,
    SitePackageVersionsElementModelFromJSONTyped,
    SitePackageVersionsElementModelToJSON,
} from './SitePackageVersionsElementModel';

/**
 * 
 * @export
 * @interface SiteSourcePackagesVersionsModel
 */
export interface SiteSourcePackagesVersionsModel {
    /**
     * request arguments
     * @type {object}
     * @memberof SiteSourcePackagesVersionsModel
     */
    requestArgs?: object;
    /**
     * Packages downloads
     * @type {Array<SitePackageVersionsElementModel>}
     * @memberof SiteSourcePackagesVersionsModel
     */
    versions?: Array<SitePackageVersionsElementModel>;
}

/**
 * Check if a given object implements the SiteSourcePackagesVersionsModel interface.
 */
export function instanceOfSiteSourcePackagesVersionsModel(value: object): value is SiteSourcePackagesVersionsModel {
    return true;
}

export function SiteSourcePackagesVersionsModelFromJSON(json: any): SiteSourcePackagesVersionsModel {
    return SiteSourcePackagesVersionsModelFromJSONTyped(json, false);
}

export function SiteSourcePackagesVersionsModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): SiteSourcePackagesVersionsModel {
    if (json == null) {
        return json;
    }
    return {
        
        'requestArgs': json['request_args'] == null ? undefined : json['request_args'],
        'versions': json['versions'] == null ? undefined : ((json['versions'] as Array<any>).map(SitePackageVersionsElementModelFromJSON)),
    };
}

export function SiteSourcePackagesVersionsModelToJSON(value?: SiteSourcePackagesVersionsModel | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'request_args': value['requestArgs'],
        'versions': value['versions'] == null ? undefined : ((value['versions'] as Array<any>).map(SitePackageVersionsElementModelToJSON)),
    };
}