// src/rdb/models/SiteFingPackagesModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ALTRepo API
 * altrepo API v1
 *
 * The version of the OpenAPI document: 1.19.15
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SiteFingPackagesPackageModel } from './SiteFingPackagesPackageModel';
import {
    SiteFingPackagesPackageModelFromJSON,
    SiteFingPackagesPackageModelFromJSONTyped,
    SiteFingPackagesPackageModelToJSON,
} from './SiteFingPackagesPackageModel';

/**
 * 
 * @export
 * @interface SiteFingPackagesModel
 */
export interface SiteFingPackagesModel {
    /**
     * request arguments
     * @type {object}
     * @memberof SiteFingPackagesModel
     */
    requestArgs?: object;
    /**
     * number of packages found
     * @type {number}
     * @memberof SiteFingPackagesModel
     */
    length?: number;
    /**
     * found packages
     * @type {Array<SiteFingPackagesPackageModel>}
     * @memberof SiteFingPackagesModel
     */
    packages?: Array<SiteFingPackagesPackageModel>;
}

/**
 * Check if a given object implements the SiteFingPackagesModel interface.
 */
export function instanceOfSiteFingPackagesModel(value: object): value is SiteFingPackagesModel {
    return true;
}

export function SiteFingPackagesModelFromJSON(json: any): SiteFingPackagesModel {
    return SiteFingPackagesModelFromJSONTyped(json, false);
}

export function SiteFingPackagesModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): SiteFingPackagesModel {
    if (json == null) {
        return json;
    }
    return {
        
        'requestArgs': json['request_args'] == null ? undefined : json['request_args'],
        'length': json['length'] == null ? undefined : json['length'],
        'packages': json['packages'] == null ? undefined : ((json['packages'] as Array<any>).map(SiteFingPackagesPackageModelFromJSON)),
    };
}

export function SiteFingPackagesModelToJSON(value?: SiteFingPackagesModel | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'request_args': value['requestArgs'],
        'length': value['length'],
        'packages': value['packages'] == null ? undefined : ((value['packages'] as Array<any>).map(SiteFingPackagesPackageModelToJSON)),
    };
}