// src/rdb/models/SitePackageVersionsElementModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ALTRepo API
 * altrepo API v1
 *
 * The version of the OpenAPI document: 1.19.15
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SitePackageVersionsElementModel
 */
export interface SitePackageVersionsElementModel {
    /**
     * package set name
     * @type {string}
     * @memberof SitePackageVersionsElementModel
     */
    branch?: string;
    /**
     * package version
     * @type {string}
     * @memberof SitePackageVersionsElementModel
     */
    version?: string;
    /**
     * package release
     * @type {string}
     * @memberof SitePackageVersionsElementModel
     */
    release?: string;
    /**
     * package hash UInt64 as string
     * @type {string}
     * @memberof SitePackageVersionsElementModel
     */
    pkghash?: string;
    /**
     * package was deleted from branch
     * @type {boolean}
     * @memberof SitePackageVersionsElementModel
     */
    deleted?: boolean;
}

/**
 * Check if a given object implements the SitePackageVersionsElementModel interface.
 */
export function instanceOfSitePackageVersionsElementModel(value: object): value is SitePackageVersionsElementModel {
    return true;
}

export function SitePackageVersionsElementModelFromJSON(json: any): SitePackageVersionsElementModel {
    return SitePackageVersionsElementModelFromJSONTyped(json, false);
}

export function SitePackageVersionsElementModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): SitePackageVersionsElementModel {
    if (json == null) {
        return json;
    }
    return {
        
        'branch': json['branch'] == null ? undefined : json['branch'],
        'version': json['version'] == null ? undefined : json['version'],
        'release': json['release'] == null ? undefined : json['release'],
        'pkghash': json['pkghash'] == null ? undefined : json['pkghash'],
        'deleted': json['deleted'] == null ? undefined : json['deleted'],
    };
}

export function SitePackageVersionsElementModelToJSON(value?: SitePackageVersionsElementModel | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'branch': value['branch'],
        'version': value['version'],
        'release': value['release'],
        'pkghash': value['pkghash'],
        'deleted': value['deleted'],
    };
}